Added function to determine song, fixed timestamp
Corrected timestamp to local time rather than universal time
Added optional string parameter to make_csv(). Enabling user to
determine the text on the first line of the file.
Added what_song() function to Interface.py to determine the song that
was just played.
Can pass what_song() to  make_csv() as optional_str to add name of song
to the beginning of the log file